staff protal implement and some fixed in backend
Staff Portal Implemented in Frontend, where only doctor can see the staffportal and cancel the appointments booked by his patients. And the api for fetching the appointments is created and fix in booking appointment such as the Status should SET to Booked when booking an appointment is fixed. And also some conditions in Upcoming appointments api is changed so that the actual data comes to the actual user .